requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-autoslug==1.9.9
django-jazzmin==2.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
promise==2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
